+++ jodag-spring/src/main/java/jodag/generator/orm/mybatis/MyBatisLoader.java
@@ -1,6 +1,5 @@
 package jodag.generator.orm.mybatis;
 
-import jodag.generator.Path;
 import jodag.generator.orm.ORMLoader;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
