+++ jodag-spring/src/main/java/jodag/generator/orm/mybatis/MyBatisMetadata.java
@@ -1,10 +1,8 @@
 package jodag.generator.orm.mybatis;
 
-import jodag.generator.Path;
 import org.springframework.core.io.Resource;
 import org.springframework.stereotype.Component;
 
-import java.lang.reflect.Field;
 import java.util.*;
 
 @Component
