+++ jodag-spring/src/main/java/jodag/generator/orm/mybatis/Path.java
@@ -1,4 +1,4 @@
-package jodag.generator;
+package jodag.generator.orm.mybatis;
 
 import java.util.Objects;
