+++ jodag-spring/src/test/java/jodag/mybatis/associations/Child.java
@@ -16,7 +16,6 @@ public class Child {
         return "Child{" +
                 "id=" + id +
                 ", name='" + name + '\'' +
-                ", parent=" + parent +
                 '}';
     }
 }
